refactor: parameters into more atomic files
